payment_plan/requirements.txt
Django==5.1.7
psycopg2-binary>=2.9,<3.0
django-cors-headers==4.4.0
django-filter==23.5
django-ratelimit==4.1.0
django-templated-mail==1.1.1
django-translations==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python_dateutil==2.9.0.post0
celery==5.5.3
django-celery-beat==2.8.1
redis>=4.0,<6.0